+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 Features:
 
   - documents returned by a query have the new score property that reflects the match relevance in percent (1-100%)
-  - added compatibility to the kaminari pagination gem (Thanks, Javi)
+  - added compatibility to the kaminari pagination gem (thanks, Javi)
+  - added support for phrase searches (XapianDb.search('"this exact sentence"'))
 
 ##0.5.5 (February 25th, 2011)